+++ compiler/cmd/unigo/main.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 	"os"
 
-	"garnet/compiler/compile"
+	"unigo/compiler/compile"
 )
 
 func main() {
 	if len(os.Args) != 3 || os.Args[1] != "build" {
-		fmt.Fprintln(os.Stderr, "用法: garnet build <file.gn>")
+		fmt.Fprintln(os.Stderr, "用法: unigo build <file.ug>")
 		os.Exit(2)
 	}
 	if err := compile.Build(os.Args[2]); err != nil {