+++ editors/tree-sitter-unigo/src/parser.c
@@ -8856,7 +8856,7 @@ extern "C" {
 #define TS_PUBLIC __attribute__((visibility("default")))
 #endif
 
-TS_PUBLIC const TSLanguage *tree_sitter_garnet(void) {
+TS_PUBLIC const TSLanguage *tree_sitter_unigo(void) {
   static const TSLanguage language = {
     .abi_version = LANGUAGE_VERSION,
     .symbol_count = SYMBOL_COUNT,
@@ -8889,7 +8889,7 @@ TS_PUBLIC const TSLanguage *tree_sitter_garnet(void) {
     .keyword_lex_fn = ts_lex_keywords,
     .keyword_capture_token = sym_identifier,
     .primary_state_ids = ts_primary_state_ids,
-    .name = "garnet",
+    .name = "unigo",
     .max_reserved_word_set_size = 0,
     .metadata = {
       .major_version = 0,